requirements.txt
websocket-client==0.26.0
requests==2.5.3
gitpython==0.3.6